requirements.txt
scipy
pydub
tensorflow-gpu
keras
matplotlib